+++ bin/reset.dart
@@ -1,0 +1,4 @@
+/// Run to reset launcher icons
+void main(List<String> arguments) {
+  print('\nReset launcher icons feature not yet implemented. 😭\n');
+}